🚀 v2.0.0: promote v2 to crate root, remove 1.x
BREAKING: v2 is now the crate root; the entire 1.x API (ChainBuilder/Client/Select/
trait-based builder) is removed. src/v2/* moved to src/; lib.rs re-exports the
typed dialect-generic API at root.

- Cargo.toml → 2.0.0; features: default=["sqlx_mysql"], + sqlx_sqlite/sqlx_postgres/
  json (dep:serde_json). Dropped serde (unused), v2/mysql/sqlite/postgres/
  dev-dependencies features. All 3 sqlx drivers now coexist (1.x dup-def gone).
- Tests: removed 1.x suites; v2_* → * (12 files), imports chain_builder::v2::→::,
  dropped #![cfg(feature="v2")]. dev-dep sqlx with all drivers for test compile.
- Docs: README + docs/guide.md rewritten for the v2 API; CHANGELOG 2.0.0.

Verified: default+doctest green, all-3-backends 144 tests, clippy -D warnings clean,
fmt clean, packages (40 files). Not published (lands on v2 branch for review).
