--- src/common/mod.rs
@@ -1,4 +1,0 @@
-pub mod join_compiler;
-pub mod method_compiler;
-pub mod operator_to_sql;
-pub mod statement_compiler;
