+++ tests/sqlx.rs
@@ -4,8 +4,6 @@
 //! one-backend feature sets (the 1.x crate cannot build with 2+ sqlx backends
 //! at once). Asserts the sqlx query's SQL equals the builder's `to_sql().0`.
 
-#![cfg(feature = "v2")]
-
 // A trivial row type used only for the `to_sqlx_query_as` compile-smoke checks.
 #[cfg(any(
     feature = "sqlx_postgres",
@@ -22,7 +20,9 @@ struct IdRow {
 impl<'r> sqlx::FromRow<'r, sqlx::postgres::PgRow> for IdRow {
     fn from_row(row: &'r sqlx::postgres::PgRow) -> Result<Self, sqlx::Error> {
         use sqlx::Row;
-        Ok(IdRow { id: row.try_get("id")? })
+        Ok(IdRow {
+            id: row.try_get("id")?,
+        })
     }
 }
 
@@ -30,7 +30,9 @@ impl<'r> sqlx::FromRow<'r, sqlx::postgres::PgRow> for IdRow {
 impl<'r> sqlx::FromRow<'r, sqlx::mysql::MySqlRow> for IdRow {
     fn from_row(row: &'r sqlx::mysql::MySqlRow) -> Result<Self, sqlx::Error> {
         use sqlx::Row;
-        Ok(IdRow { id: row.try_get("id")? })
+        Ok(IdRow {
+            id: row.try_get("id")?,
+        })
     }
 }
 
@@ -38,14 +40,16 @@ impl<'r> sqlx::FromRow<'r, sqlx::mysql::MySqlRow> for IdRow {
 impl<'r> sqlx::FromRow<'r, sqlx::sqlite::SqliteRow> for IdRow {
     fn from_row(row: &'r sqlx::sqlite::SqliteRow) -> Result<Self, sqlx::Error> {
         use sqlx::Row;
-        Ok(IdRow { id: row.try_get("id")? })
+        Ok(IdRow {
+            id: row.try_get("id")?,
+        })
     }
 }
 
 #[cfg(feature = "sqlx_postgres")]
 #[test]
 fn postgres_to_sqlx_sql_matches_to_sql() {
-    use chain_builder::v2::{Postgres, QueryBuilder};
+    use chain_builder::{Postgres, QueryBuilder};
     use sqlx::Execute;
     let qb = QueryBuilder::<Postgres>::table("users")
         .select(["id"])
@@ -58,7 +62,7 @@ fn postgres_to_sqlx_sql_matches_to_sql() {
 #[cfg(feature = "sqlx_postgres")]
 #[test]
 fn postgres_to_sqlx_query_as_compiles() {
-    use chain_builder::v2::{Postgres, QueryBuilder};
+    use chain_builder::{Postgres, QueryBuilder};
     use sqlx::Execute;
     let qb = QueryBuilder::<Postgres>::table("users")
         .select(["id"])
@@ -71,7 +75,7 @@ fn postgres_to_sqlx_query_as_compiles() {
 #[cfg(feature = "sqlx_mysql")]
 #[test]
 fn mysql_to_sqlx_sql_matches_to_sql() {
-    use chain_builder::v2::{MySql, QueryBuilder};
+    use chain_builder::{MySql, QueryBuilder};
     use sqlx::Execute;
     let qb = QueryBuilder::<MySql>::table("users")
         .select(["id"])
@@ -84,7 +88,7 @@ fn mysql_to_sqlx_sql_matches_to_sql() {
 #[cfg(feature = "sqlx_mysql")]
 #[test]
 fn mysql_to_sqlx_query_as_compiles() {
-    use chain_builder::v2::{MySql, QueryBuilder};
+    use chain_builder::{MySql, QueryBuilder};
     use sqlx::Execute;
     let qb = QueryBuilder::<MySql>::table("users")
         .select(["id"])
@@ -97,7 +101,7 @@ fn mysql_to_sqlx_query_as_compiles() {
 #[cfg(feature = "sqlx_sqlite")]
 #[test]
 fn sqlite_to_sqlx_sql_matches_to_sql() {
-    use chain_builder::v2::{QueryBuilder, Sqlite};
+    use chain_builder::{QueryBuilder, Sqlite};
     use sqlx::Execute;
     let qb = QueryBuilder::<Sqlite>::table("users")
         .select(["id"])
@@ -110,7 +114,7 @@ fn sqlite_to_sqlx_sql_matches_to_sql() {
 #[cfg(feature = "sqlx_sqlite")]
 #[test]
 fn sqlite_to_sqlx_query_as_compiles() {
-    use chain_builder::v2::{QueryBuilder, Sqlite};
+    use chain_builder::{QueryBuilder, Sqlite};
     use sqlx::Execute;
     let qb = QueryBuilder::<Sqlite>::table("users")
         .select(["id"])
